fix: address PR review feedback
- CodexExtras: cross-platform shortPath regex (macOS, Linux, Windows)
- Extract deepMerge to shared utils.ts, remove duplication from gemini/codex
- Add .codex/ to project .gitignore during session to prevent leaking local paths
- Case-insensitive INTERNAL_BLOCKED check in codex.ts and gemini.ts (Windows safety)
- Add codex-compat workflow, prompt, and version tracker
