+++ .github/codex-version.json
@@ -1,0 +1,4 @@
+{
+  "lastCheckedVersion": "rust-v0.118.0",
+  "lastCheckedAt": "2026-04-10T00:00:00Z"
+}
